requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-rest==0.8.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
DoubleMetaphone==1.1
idna==3.3
jellyfish==0.9.0
psycopg2==2.9.5
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
